refactor: change actions name and moved the constants to command folder
