+++ src/store/actions/index.js
@@ -1,2 +1,2 @@
-export { default as userPostFetch } from './userPostFetch';
+export { default as userLogIn } from './userLogIn';
 export { default as userLogOut } from './userLogOut';
